+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - EntityPanel.selectDetailPanel() removed for now.
 - EntityPanel.containsDetailPanel() removed.
 - EntityPanel.detailPanel(entityType) refactored, error message improved.
+- EntityPanel bug fixed, editControlTablePanel was not created in case of no table panel, renamed mainPanel, tests added.
 
 ## 0.17.33
 ### is.codion.common.core